flask-api/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.0
mysql-connector-python==8.0.33
requests==2.31.0
